Revert "futex-multiple-wait" patch series from master
Due to the iteration speed of this patch series, lets just revert the
branch(es) we merged earlier and keep each version in its own branch.
This will keep the feature history itself clean, and make it more clear
that we're updating to a new version.

Latest version can be found here:
https://gitlab.collabora.com/tonyk/linux/commits/futex-dev-v3

A prerequisite backport from upstream is required first (for 5.5):
202bf8d7 - "compat: provide compat_ptr() on all architectures"