Revert "mm: Proactive compaction"
Unfortunately, this brings back the poor desktop responsiveness we get
with watermark boosting that's already built-in to mainline.  Disabling
watermark boosting and reducing unnecessary activations of kswapd really
helps with keeping the experience of using the desktop as consistent as
possible.

I was hoping this patch would find a way of doing compaction without
interfering with desktop class applications, but unfortunately it also
causes important applications like Xorg and anything on top of it to
freeze.

This reverts commit b137d23d28ed58a9ca30a30297b1875cb689542d.