mm: Don't stop kswapd on a per-node basis when there are no waiters
The page allocator wakes all kswapds in an allocation context's allowed
nodemask in the slow path, so it doesn't make sense to have the kswapd-
waiter count per each NUMA node. Instead, it should be a global counter
to stop all kswapds when there are no failed allocation requests.

Conflicts:
	mm/page_alloc.c

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: Simão Gomes Viana <simao@halogenos.org>