ZEN: Reduce module compression level for ZSTD
Using the original values of "-T0 -20 --ultra" leads to compression runtime
of over a minute and a half on a 374m module directory, reducing folder
size to 108m.  This is using a Haswell CPU running at 4.2ghz on a single
core.

Dropping compress levels down to "-T0 -9" increases the final compressed
folder size by 4m but reduces the total runtime of compression to 6
seconds.  When penny pinching with ondemand cloud CI, this tradeoff is
completely worthwhile.

Signed-off-by: Simão Gomes Viana <simao@halogenos.org>