md/raid6 algorithms: scale test duration for speedier boots
The original code runs for a set run time based on the duration of
2 ^ RAID6_TIME_JIFFIES_LG2. The default kernel value for
RAID6_TIME_JIFFIES_LG2 is 4, however, emperical testing shows that a
value of 3.5 for RAID6_TIME_JIFFIES_LG2 is the sweet spot for getting
consistent benchmarking results and reducing up the calibration run time
by 29.4%.

To achieve 2^3.5 we use the following:
   2^3.5 = 2^4 / 2^0.5
         = 2^4 / sqrt(2)
         = 2^4 * 0.707106781

Too keep this as integer math that is as accurate as required and avoiding
overflow, this becomes:
         = 2^4 * 181 / 256
         = (2^4 * 181) >> 8

We also need to scale down perf by the same factor, however, to
get a good approximate integer result without an overflow we scale
by 2^4.0 * sqrt(2) =
         = 2 ^ 4 * 1.41421356237
         = 2 ^ 4 * 1448 / 1024
         = (2 ^ 4 * 1448) >> 10

Tested on an Alderlake i9-12900, reduced raid6 calibration on boot
from 68000 microseconds to 48000 microseconds, saving 0.02 seconds.
Testing by booting 5 times and double checking the algorithm
chosen is the same as pre-optimised version.

Originally used in Ubuntu (written by Colin Ian King) and forward ported
to 6.1 kernels.

Signed-off-by: Colin Ian King <colin.king@intel.com>
Signed-off-by: Simão Gomes Viana <simao@halogenos.org>
Signed-off-by: Simão Gomes Viana <devel@superboring.dev>